requirements.txt
bidict
pyserial